Move harn test dispatch into test command module
