+++ crates/harn-cli/src/lib.rs
@@ -763,252 +763,7 @@ async fn async_main(raw_args: Vec<String>, runtime_mode: CliRuntimeMode) {
                 commands::check::fmt_targets(&targets, mode, &opts);
             }
         }
-        Command::Test(args) => {
-            if args.watch && (args.junit.is_some() || args.json_out.is_some()) {
-                command_error(
-                    "`harn test --watch` cannot combine with --junit or --json-out; the watch loop never terminates so the report would never be written",
-                );
-            }
-            let shard_requested = args.shard_index.is_some() || args.shard_total.is_some();
-            if args.target.as_deref() == Some("agents-conformance") {
-                if args.selection.is_some() {
-                    command_error(
-                        "`harn test agents-conformance` does not accept a second positional target; use --category instead",
-                    );
-                }
-                if args.evals
-                    || args.determinism
-                    || args.record
-                    || args.replay
-                    || args.watch
-                    || shard_requested
-                {
-                    command_error(
-                        "`harn test agents-conformance` cannot be combined with --evals, --determinism, --record, --replay, --watch, or test sharding",
-                    );
-                }
-                let Some(target_url) = args.agents_target.clone() else {
-                    command_error("`harn test agents-conformance` requires --target <url>");
-                };
-                commands::agents_conformance::run_agents_conformance(
-                    commands::agents_conformance::AgentsConformanceConfig {
-                        target_url,
-                        api_key: args.agents_api_key.clone(),
-                        categories: args.agents_category.clone(),
-                        timeout_ms: args.timeout,
-                        verbose: args.verbose,
-                        json: args.json,
-                        json_out: args.json_out.clone(),
-                        workspace_id: args.agents_workspace_id.clone(),
-                        session_id: args.agents_session_id.clone(),
-                    },
-                )
-                .await;
-                return;
-            }
-            if args.target.as_deref() == Some("protocols") {
-                if args.evals || args.determinism || args.record || args.replay || args.watch {
-                    command_error(
-                        "`harn test protocols` cannot be combined with --evals, --determinism, --record, --replay, or --watch",
-                    );
-                }
-                if args.junit.is_some()
-                    || args.agents_target.is_some()
-                    || args.agents_api_key.is_some()
-                    || !args.agents_category.is_empty()
-                    || args.json
-                    || args.json_out.is_some()
-                    || args.agents_workspace_id.is_some()
-                    || args.agents_session_id.is_some()
-                    || args.parallel
-                    || shard_requested
-                    || !args.skill_dir.is_empty()
-                {
-                    command_error(
-                        "`harn test protocols` accepts only --filter, --verbose, --timing, and an optional fixture selection",
-                    );
-                }
-                commands::protocol_conformance::run_protocol_conformance(
-                    args.selection.as_deref(),
-                    args.filter.as_deref(),
-                    args.verbose || args.timing,
-                );
-                return;
-            }
-            if args.evals {
-                if shard_requested {
-                    command_error("--evals cannot be combined with test sharding");
-                }
-                if args.determinism || args.record || args.replay || args.watch {
-                    command_error("--evals cannot be combined with --determinism, --record, --replay, or --watch");
-                }
-                if args.target.as_deref() != Some("package") || args.selection.is_some() {
-                    command_error("package evals are run with `harn test package --evals`");
-                }
-                run_package_evals();
-            } else if args.determinism {
-                if shard_requested {
-                    command_error("--determinism cannot be combined with test sharding");
-                }
-                let cli_skill_dirs: Vec<PathBuf> =
-                    args.skill_dir.iter().map(PathBuf::from).collect();
-                if args.watch {
-                    command_error("--determinism cannot be combined with --watch");
-                }
-                if args.record || args.replay {
-                    command_error("--determinism manages its own record/replay cycle");
-                }
-                if let Some(t) = args.target.as_deref() {
-                    if t == "conformance" {
-                        commands::test::run_conformance_determinism_tests(
-                            t,
-                            args.selection.as_deref(),
-                            args.filter.as_deref(),
-                            args.timeout,
-                            &cli_skill_dirs,
-                        )
-                        .await;
-                    } else if args.selection.is_some() {
-                        command_error(
-                            "only `harn test conformance` accepts a second positional target",
-                        );
-                    } else {
-                        commands::test::run_determinism_tests(
-                            t,
-                            args.filter.as_deref(),
-                            args.timeout,
-                            &cli_skill_dirs,
-                        )
-                        .await;
-                    }
-                } else {
-                    let test_dir = if PathBuf::from("tests").is_dir() {
-                        "tests".to_string()
-                    } else {
-                        command_error("no path specified and no tests/ directory found");
-                    };
-                    if args.selection.is_some() {
-                        command_error(
-                            "only `harn test conformance` accepts a second positional target",
-                        );
-                    }
-                    commands::test::run_determinism_tests(
-                        &test_dir,
-                        args.filter.as_deref(),
-                        args.timeout,
-                        &cli_skill_dirs,
-                    )
-                    .await;
-                }
-            } else {
-                let cli_skill_dirs: Vec<PathBuf> =
-                    args.skill_dir.iter().map(PathBuf::from).collect();
-                if args.record {
-                    harn_vm::llm::set_replay_mode(
-                        harn_vm::llm::LlmReplayMode::Record,
-                        ".harn-fixtures",
-                    );
-                } else if args.replay {
-                    harn_vm::llm::set_replay_mode(
-                        harn_vm::llm::LlmReplayMode::Replay,
-                        ".harn-fixtures",
-                    );
-                }
-
-                if let Some(t) = args.target.as_deref() {
-                    if t == "conformance" {
-                        if shard_requested {
-                            command_error("test sharding is only supported for user test suites");
-                        }
-                        commands::test::run_conformance_tests(
-                            t,
-                            args.selection.as_deref(),
-                            args.filter.as_deref(),
-                            args.junit.as_deref(),
-                            args.timeout,
-                            commands::test::ConformanceRunOptions {
-                                verbose: args.verbose,
-                                timing: args.timing,
-                                differential_optimizations: args.differential_optimizations,
-                                json: args.json,
-                                cli_skill_dirs: &cli_skill_dirs,
-                            },
-                        )
-                        .await;
-                    } else if args.selection.is_some() {
-                        command_error(
-                            "only `harn test conformance` accepts a second positional target",
-                        );
-                    } else {
-                        let shard = resolve_user_test_shard(args.shard_index, args.shard_total);
-                        let run_args = commands::test::UserTestRunArgs {
-                            filter: args.filter.as_deref(),
-                            timeout_ms: args.timeout,
-                            max_test_ms: args.max_test_ms,
-                            max_execute_ms: args.max_execute_ms,
-                            parallel: args.parallel,
-                            jobs: args.jobs,
-                            shard,
-                            verbose: args.verbose,
-                            timing: args.timing,
-                            diagnose: args.diagnose,
-                            cli_skill_dirs: &cli_skill_dirs,
-                        };
-                        if args.watch {
-                            commands::test::run_watch_tests(t, run_args).await;
-                        } else {
-                            commands::test::run_user_tests(
-                                t,
-                                run_args,
-                                commands::test::UserTestReportConfig {
-                                    junit_path: args.junit.as_deref(),
-                                    json_out_path: args.json_out.as_deref(),
-                                },
-                            )
-                            .await;
-                        }
-                    }
-                } else {
-                    let test_dir = if PathBuf::from("tests").is_dir() {
-                        "tests".to_string()
-                    } else {
-                        command_error("no path specified and no tests/ directory found");
-                    };
-                    if args.selection.is_some() {
-                        command_error(
-                            "only `harn test conformance` accepts a second positional target",
-                        );
-                    }
-                    let shard = resolve_user_test_shard(args.shard_index, args.shard_total);
-                    let run_args = commands::test::UserTestRunArgs {
-                        filter: args.filter.as_deref(),
-                        timeout_ms: args.timeout,
-                        max_test_ms: args.max_test_ms,
-                        max_execute_ms: args.max_execute_ms,
-                        parallel: args.parallel,
-                        jobs: args.jobs,
-                        shard,
-                        verbose: args.verbose,
-                        timing: args.timing,
-                        diagnose: args.diagnose,
-                        cli_skill_dirs: &cli_skill_dirs,
-                    };
-                    if args.watch {
-                        commands::test::run_watch_tests(&test_dir, run_args).await;
-                    } else {
-                        commands::test::run_user_tests(
-                            &test_dir,
-                            run_args,
-                            commands::test::UserTestReportConfig {
-                                junit_path: args.junit.as_deref(),
-                                json_out_path: args.json_out.as_deref(),
-                            },
-                        )
-                        .await;
-                    }
-                }
-            }
-        }
+        Command::Test(args) => commands::test::run_command(args).await,
         Command::Init(args) => {
             commands::init::init_project(args.name.as_deref(), args.template).await;
         }
@@ -2717,7 +2472,7 @@ fn print_persona_ladder_report(report: &harn_vm::orchestration::PersonaEvalLadde
     }
 }
 
-fn run_package_evals() {
+pub(crate) fn run_package_evals() {
     let paths = package::load_package_eval_pack_paths(None).unwrap_or_else(|error| {
         eprintln!("{error}");
         process::exit(1);
@@ -3320,20 +3075,6 @@ fn connector_secret_namespace(base_dir: &Path) -> String {
                 .unwrap_or("workspace");
             format!("harn/{leaf}")
         }
-    }
-}
-
-fn resolve_user_test_shard(
-    shard_index: Option<usize>,
-    shard_total: Option<usize>,
-) -> Option<test_runner::TestShard> {
-    match (shard_index, shard_total) {
-        (None, None) => None,
-        (Some(index), Some(total)) => match test_runner::TestShard::new(index, total) {
-            Ok(shard) => Some(shard),
-            Err(error) => command_error(&error),
-        },
-        _ => command_error("test sharding requires both --shard-index and --shard-total"),
     }
 }
